feat: 주문내역 뷰모델 및 뷰 구성

+++ EzyBook/Data/DTO/Review/ReviewResponseDTO.swift
@@ -75,5 +75,5 @@ struct ReviewRatingResponseDTO: Decodable {
 
 struct ReviewInfoResponseDTO: Decodable {
     let id: String
-    let rating: Int
+    let rating: Double
 }
